test(httproute): 添加mock测试数据包，服务http注册单元测试
